Change temporary data location to use XDG compliant folders. Proxy clips, timeline preview and thumbs will now be saved inside a project specific folder in $HOME/.cache/kdenlive
Ref: T2582

+++ src/timeline/customtrackview.cpp
@@ -7114,21 +7114,22 @@ void CustomTrackView::slotUpdateAllThumbs()
     QList<QGraphicsItem *> itemList = items();
     //if (itemList.isEmpty()) return;
     ClipItem *item;
-    const QString thumbBase = m_document->projectFolder().path() + "/thumbs/";
+    bool ok = false;
+    QDir thumbsFolder = m_document->getCacheDir(CacheThumbs, &ok);
     for (int i = 0; i < itemList.count(); ++i) {
         if (itemList.at(i)->type() == AVWidget) {
             item = static_cast <ClipItem *>(itemList.at(i));
             if (item && item->isEnabled() && item->clipType() != Color && item->clipType() != Audio) {
                 // Check if we have a cached thumbnail
                 if (item->clipType() == Image || item->clipType() == Text) {
-                    QString thumb = thumbBase + item->getBinHash() + "#0.png";
-                    if (QFile::exists(thumb)) {
+                    QString thumb = thumbsFolder.absoluteFilePath(item->getBinHash() + "#0.png");
+                    if (ok && QFile::exists(thumb)) {
                         QPixmap pix(thumb);
                         if (pix.isNull()) QFile::remove(thumb);
                         item->slotSetStartThumb(pix);
                     }
                 } else {
-                    QString startThumb = thumbBase + item->getBinHash() + '#';
+                    QString startThumb = thumbsFolder.absoluteFilePath(item->getBinHash() + "#");
                     QString endThumb = startThumb;
                     startThumb.append(QString::number((int) item->speedIndependantCropStart().frames(m_document->fps())) + ".png");
                     endThumb.append(QString::number((int) (item->speedIndependantCropStart() + item->speedIndependantCropDuration()).frames(m_document->fps()) - 1) + ".png");
@@ -7154,6 +7155,8 @@ void CustomTrackView::saveThumbnails()
 {
     QList<QGraphicsItem *> itemList = items();
     ClipItem *item;
+    bool ok = false;
+    QDir thumbsFolder = m_document->getCacheDir(CacheThumbs, &ok);
     QString thumbBase = m_document->projectFolder().path() + "/thumbs/";
     for (int i = 0; i < itemList.count(); ++i) {
         if (itemList.at(i)->type() == AVWidget) {
@@ -7161,13 +7164,13 @@ void CustomTrackView::saveThumbnails()
             if (item->clipType() != Color) {
                 // Check if we have a cached thumbnail
                 if (item->clipType() == Image || item->clipType() == Text || item->clipType() == Audio) {
-                    QString thumb = thumbBase + item->getBinHash() + "#0.png";
+                    QString thumb = thumbsFolder.absoluteFilePath(item->getBinHash() + "#0.png");
                     if (!QFile::exists(thumb)) {
                         QPixmap pix(item->startThumb());
                         pix.save(thumb);
                     }
                 } else {
-                    QString startThumb = thumbBase + item->getBinHash() + '#';
+                    QString startThumb = thumbsFolder.absoluteFilePath(item->getBinHash() + '#');
                     QString endThumb = startThumb;
                     startThumb.append(QString::number((int) item->speedIndependantCropStart().frames(m_document->fps())) + ".png");
                     endThumb.append(QString::number((int) (item->speedIndependantCropStart() + item->speedIndependantCropDuration()).frames(m_document->fps()) - 1) + ".png");
